+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Sun Mar 26 12:01:48 PDT 2017"
+buildTime = "Sun Mar 26 13:05:06 PDT 2017"